remove unsed generator apis
