+++ legend-engine-xts-avro/legend-engine-xt-avro-api/pom.xml
@@ -46,6 +46,14 @@
         </dependency>
         <!-- ENGINE -->
 
+        <!-- JACKSON -->
+        <dependency>
+            <groupId>com.fasterxml.jackson.core</groupId>
+            <artifactId>jackson-databind</artifactId>
+
+        </dependency>
+        <!-- JACKSON -->
+
         <!-- ANNOTATIONS -->
         <dependency>
             <groupId>io.swagger</groupId>
@@ -92,6 +100,28 @@
             <artifactId>pac4j-core</artifactId>
         </dependency>
         <!-- PAC4J -->
+
+        <!-- TEST -->
+        <dependency>
+            <groupId>junit</groupId>
+            <artifactId>junit</artifactId>
+            <scope>test</scope>
+        </dependency>
+        <dependency>
+            <groupId>org.finos.legend.engine</groupId>
+            <artifactId>legend-engine-language-pure-compiler</artifactId>
+            <type>test-jar</type>
+            <scope>test</scope>
+        </dependency>
+        <dependency>
+            <groupId>org.glassfish.jersey.core</groupId>
+            <artifactId>jersey-common</artifactId>
+            <version>${jersey.version}</version>
+            <scope>test</scope>
+        </dependency>
+
+        <!-- TEST -->
+
     </dependencies>
