+++ legend-engine-xts-avro/legend-engine-xt-avro-api/src/test/resources/org/finos/legend/engine/external/format/avro/schema/generations/api/avroInputNoConfig.json
