+++ legend-engine-xts-generation/legend-engine-external-shared/src/main/java/org/finos/legend/engine/external/shared/format/extension/GenerationExtension.java
@@ -15,9 +15,7 @@
 package org.finos.legend.engine.external.shared.format.extension;
 
 import org.finos.legend.engine.external.shared.format.generations.description.GenerationConfigurationDescription;
-import org.finos.legend.engine.external.shared.format.imports.description.ImportConfigurationDescription;
 import org.finos.legend.engine.language.pure.compiler.toPureGraph.CompileContext;
-import org.finos.legend.engine.language.pure.modelManager.ModelManager;
 import org.finos.legend.engine.protocol.pure.v1.model.packageableElement.PackageableElement;
 import org.finos.legend.engine.shared.core.extension.LegendGenerationExtension;
 import org.finos.legend.pure.generated.Root_meta_pure_generation_metamodel_GenerationConfiguration;
@@ -41,20 +39,7 @@ public interface GenerationExtension extends LegendGenerationExtension
 
     GenerationConfigurationDescription getGenerationDescription();
 
-    /**
-     * TO REMOVE
-     */
-    default ImportConfigurationDescription getImportDescription()
-    {
-        return null;
-    }
-
     Root_meta_pure_generation_metamodel_GenerationConfiguration defaultConfig(CompileContext context);
-
-    default Object getService(ModelManager modelManager)
-    {
-        return null;
-    }
 
     List<Root_meta_pure_generation_metamodel_GenerationOutput> generateFromElement(PackageableElement element, CompileContext compileContext);
 }
