+++ legend-engine-xts-json/legend-engine-xt-json-api/src/test/resources/org/finos/legend/engine/external/format/jsonSchema/schema/generations/api/simpleFileGeneration.json
